#37 refactor: change subscribeAdd -> subscribeCustomAdd
