--- Ppap/app/src/main/java/com/jeongg/ppap/presentation/subscribe_add/SubscribeAddEvent.kt
@@ -1,7 +1,0 @@
-package com.jeongg.ppap.presentation.subscribe_add
-
-sealed class SubscribeAddEvent{
-    data class EnteredTitle(val title: String): SubscribeAddEvent()
-    data class EnteredRssLink(val rss: String): SubscribeAddEvent()
-    object SaveSubscribe: SubscribeAddEvent()
-}
